requirements.txt
Django==3.1.3
django-bootstrap4==2.3.1
django-celery-beat==2.2.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-grappelli==2.14.3
django-oss-storage==1.1.1
django-python3-ldap==0.11.4
django-redis==4.12.1
django-registration-redux==2.9
django-simpleui==2021.3
django-smart-selects==1.5.9
django-timezone-field==4.1.1
djangorestframework==3.12.2
